requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-allauth
dj-rest-auth
psycopg2-binary
django-redis
celery
redis
django-celery-beat
django-cors-headers
django-filter
drf-spectacular
Pillow
django-storages
boto3
django-environ
whitenoise
gunicorn
sentry-sdk
bleach
pytest
pytest-django
factory-boy
requests
cryptography